backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
werkzeug==2.0.3
supabase==2.13.0
numpy>=1.25.0
pandas>=1.5.0
scikit-learn>=1.0.0
matplotlib>=3.4.3
seaborn>=0.11.2
requests==2.28.2
pytest==7.3.1
gunicorn==20.1.0
python-jose==3.3.0
# Remove tensorflow and pytorch for now as they're large and might be causing issues
# tensorflow==2.8.0
# torch==1.10.0 
# librosa==0.8.1
# twilio==7.0.0
pytest-flask==1.2.0

# LLM and AI dependencies
openai>=1.0.0  # OpenAI API
anthropic>=0.5.0  # Anthropic API
tiktoken>=0.5.0  # Token counting
tenacity>=8.2.0  # Retry logic
jsonschema>=4.0.0  # JSON validation 